requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pymongo==4.6.0
ollama==0.1.0
pillow==10.1.0
python-magic==0.4.27
pydub==0.25.1
pytesseract==0.3.10
numpy==1.26.2
scikit-learn==1.3.2
torch==2.1.1
pytorch-lightning==2.1.2
watchdog==3.0.0
aiofiles==23.2.1
httpx==0.25.2
requests==2.31.0
colorama==0.4.6
loguru==0.7.2
tenacity==8.2.3
tqdm==4.66.1
attrs==23.1.0
click==8.1.7
typer==0.9.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
isort==5.13.2
pre-commit==3.5.0
